+++ diff-containers/CHANGELOG.md
@@ -1,3 +1,9 @@
+## next version
+
+### Patch
+
+* Fix laziness in the `Functor` instance for `Diff`.
+
 ## 1.0.1.0 — 2023-05-11
 
 * Add `numInserts` and `numDeletes` to count a diff's number of inserts and