+++ diff-containers/CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Patch
 
 * Fix laziness in the `Functor` instance for `Diff`.
+* Make buildable with `ghc-9.4` and `ghc-9.6`.
 
 ## 1.0.1.0 — 2023-05-11