+++ diff-containers/CHANGELOG.md
@@ -1,4 +1,4 @@
-## next version
+## 1.1.0.0 — 2023-07-14
 
 ### Breaking changes